docs(load-scores): plan complet SM/SN (inventaire TSS) + migration p5s numeric
- PROMPT_SM_SN_COMPLET.md : inventaire complet des usages TSS (calcul/lecture/affichage),
  données vérifiées en base, architecture du moteur SM/SN, statut de livraison honnête.
- add_load_scores.sql : p5s_watts en numeric (conforme au spec). Migration non encore appliquée.
- Moteur SM/SN, hooks, PMC dual et remplacements UI : en cours (pas encore dans ce commit).

+++ src/supabase/migrations/add_load_scores.sql
@@ -15,7 +15,7 @@ COMMENT ON COLUMN activities.sn_score IS 'Score neuromusculaire 0-100 (mécaniqu
 
 -- Benchmarks manquants pour les scores (puissance 5 s vélo + 1RM muscu par exercice).
 ALTER TABLE athlete_performance_profile
-  ADD COLUMN IF NOT EXISTS p5s_watts        integer,
+  ADD COLUMN IF NOT EXISTS p5s_watts        numeric,
   ADD COLUMN IF NOT EXISTS one_rm_estimates jsonb;
 
 COMMENT ON COLUMN athlete_performance_profile.p5s_watts IS 'Puissance max 5 secondes (W) — score neuromusculaire vélo';
